Update HelpAltSearch_en.html
Lines aligement with the Italian version. Many corrections.
This is the preliminary work to the discussion about what to do with lines containing references to Ooo vs LibreOffice and to the necessity of upate some links.
I'll discuss these things tomorrow or maybe the day after tomorrow in the Issue section.